agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["CopilotKit"]
readme = "README.md"
packages = [{include = "sample_agent"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
ag-ui-langgraph = "0.0.18a0"
langchain = "0.3.26"
langchain-openai = "0.3.28"
langgraph = "0.4.10"
langsmith = "0.4.4"
openai = "^1.68.2"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.0"
websockets = "^13.0"

[tool.poetry.scripts]
demo = "sample_agent.demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
